WEB-4460 rename to PatientLastReviewed Generic Adapter
