+++ app/pages/clinicworkspace/TideDashboardV2/PatientLastReviewed.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import PatientLastReviewedGenericAdapter from '../components/ReviewPatientToggle/PatientLastReviewedGenericAdapter';
+
+const PatientLastReviewed = (props) => {
+  return <PatientLastReviewedGenericAdapter {...props} />;
+};
+
+export default PatientLastReviewed;
