+++ app/pages/dashboard/PatientDrawer/MenuBar/MenuBar.js
@@ -8,7 +8,7 @@ import { Flex, Box, Text } from 'theme-ui';
 import { colors as vizColors } from '@tidepool/viz';
 import Button from '../../../../components/elements/Button';
 import moment from 'moment';
-import PatientLastReviewed from '../../../clinicworkspace/components/ReviewPatientToggle/PatientLastReviewed';
+import PatientLastReviewed from '../../../clinicworkspace/components/ReviewPatientToggle/PatientLastReviewedGenericAdapter';
 import CGMClipboardButton from './CGMClipboardButton';
 import api from '../../../../core/api';
 import { map, keys } from 'lodash';
